feat(benchmark): minor refactor
