+++ alchemist-engine/src/main/java/it/unibo/alchemist/core/model/OutputReplayStrategy.kt
@@ -7,7 +7,7 @@
  * as described in the file LICENSE in the Alchemist distribution's top directory.
  */
 
-package it.unibo.alchemist.model
+package it.unibo.alchemist.core.model
 
 /**
  * Engine output replay strategy modes.
